# examples/onk-simple-api/api/server.gen.go
// Code generated by onek. DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"net/http"
	"strconv"
)

func writeJSONError(w http.ResponseWriter, status int, message string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(map[string]string{"message": message})
}

func parseInt32(s string) int32     { v, _ := strconv.ParseInt(s, 10, 32); return int32(v) }
func parseInt64(s string) int64     { v, _ := strconv.ParseInt(s, 10, 64); return v }
func parseUint32(s string) uint32   { v, _ := strconv.ParseUint(s, 10, 32); return uint32(v) }
func parseUint64(s string) uint64   { v, _ := strconv.ParseUint(s, 10, 64); return v }
func parseFloat32(s string) float32 { v, _ := strconv.ParseFloat(s, 32); return float32(v) }
func parseFloat64(s string) float64 { v, _ := strconv.ParseFloat(s, 64); return v }
func parseBool(s string) bool       { v, _ := strconv.ParseBool(s); return v }

type UserServiceServer interface {
	CreateUser(ctx context.Context, req *CreateUserRequest) (*User, error)
	GetUser(ctx context.Context, req *GetUserRequest) (*User, error)
	Login(ctx context.Context, req *LoginRequest) (*LoginResponse, error)
}

func RegisterUserServiceServer(mux *http.ServeMux, srv UserServiceServer) {
	mux.HandleFunc("POST /api/v1/users", func(w http.ResponseWriter, r *http.Request) {
		req := new(CreateUserRequest)
		if r.Body != nil {
			if err := json.NewDecoder(r.Body).Decode(req); err != nil && err.Error() != "EOF" {
				writeJSONError(w, http.StatusBadRequest, "invalid request body: "+err.Error())
				return
			}
		}
		if r.Header.Get("X-API-Key") == "" {
			writeJSONError(w, http.StatusBadRequest, "missing required header: X-API-Key")
			return
		}
		if v, ok := any(req).(interface{ Validate() error }); ok {
			if err := v.Validate(); err != nil {
				writeJSONError(w, http.StatusBadRequest, err.Error())
				return
			}
		}
		resp, err := srv.CreateUser(r.Context(), req)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, err.Error())
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	})
	mux.HandleFunc("POST /api/v1/users/get", func(w http.ResponseWriter, r *http.Request) {
		req := new(GetUserRequest)
		if r.Body != nil {
			if err := json.NewDecoder(r.Body).Decode(req); err != nil && err.Error() != "EOF" {
				writeJSONError(w, http.StatusBadRequest, "invalid request body: "+err.Error())
				return
			}
		}
		if r.Header.Get("X-API-Key") == "" {
			writeJSONError(w, http.StatusBadRequest, "missing required header: X-API-Key")
			return
		}
		if v, ok := any(req).(interface{ Validate() error }); ok {
			if err := v.Validate(); err != nil {
				writeJSONError(w, http.StatusBadRequest, err.Error())
				return
			}
		}
		resp, err := srv.GetUser(r.Context(), req)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, err.Error())
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	})
	mux.HandleFunc("POST /api/v1/auth/login", func(w http.ResponseWriter, r *http.Request) {
		req := new(LoginRequest)
		if r.Body != nil {
			if err := json.NewDecoder(r.Body).Decode(req); err != nil && err.Error() != "EOF" {
				writeJSONError(w, http.StatusBadRequest, "invalid request body: "+err.Error())
				return
			}
		}
		if r.Header.Get("X-API-Key") == "" {
			writeJSONError(w, http.StatusBadRequest, "missing required header: X-API-Key")
			return
		}
		if r.Header.Get("X-Request-Id") == "" {
			writeJSONError(w, http.StatusBadRequest, "missing required header: X-Request-Id")
			return
		}
		if v, ok := any(req).(interface{ Validate() error }); ok {
			if err := v.Validate(); err != nil {
				writeJSONError(w, http.StatusBadRequest, err.Error())
				return
			}
		}
		resp, err := srv.Login(r.Context(), req)
		if err != nil {
			writeJSONError(w, http.StatusInternalServerError, err.Error())
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_ = json.NewEncoder(w).Encode(resp)
	})
}