# examples/onk-simple-api/api/types.gen.go
// Code generated by onek. DO NOT EDIT.
package api

import (
	"encoding/json"
)

type User struct {
	Id        string `json:"id,omitempty"`
	Name      string `json:"name,omitempty"`
	Email     string `json:"email,omitempty"`
	CreatedAt int64  `json:"created_at,omitempty"`
}

type CreateUserRequest struct {
	Name  string `json:"name,omitempty"`
	Email string `json:"email,omitempty"`
}

type GetUserRequest struct {
	Id string `json:"id,omitempty"`
}

type EmailAuth struct {
	Email         string `json:"email,omitempty"`
	Password      string `json:"password,omitempty"`
	UseTwoFactor  bool   `json:"use_two_factor,omitempty"`
	TwoFactorCode string `json:"two_factor_code,omitempty"`
}

type TokenAuth struct {
	Token     string `json:"token,omitempty"`
	TokenType string `json:"token_type,omitempty"`
	IssuedAt  int64  `json:"issued_at,omitempty"`
}

type SocialAuth struct {
	Provider    string   `json:"provider,omitempty"`
	AccessToken string   `json:"access_token,omitempty"`
	RedirectUri string   `json:"redirect_uri,omitempty"`
	State       string   `json:"state,omitempty"`
	Permissions []string `json:"permissions,omitempty"`
}

type LoginRequest struct {
	DeviceId   string                 `json:"device_id,omitempty"`
	IpAddress  string                 `json:"ip_address,omitempty"`
	RetryCount int32                  `json:"retry_count,omitempty"`
	UserAgent  string                 `json:"user_agent,omitempty"`
	AuthMethod LoginRequestAuthMethod `json:"auth_method,omitempty"`
	RememberMe bool                   `json:"remember_me,omitempty"`
	SessionId  string                 `json:"session_id,omitempty"`
	Timestamp  int64                  `json:"timestamp,omitempty"`
	Scopes     []string               `json:"scopes,omitempty"`
}

func (m *LoginRequest) MarshalJSON() ([]byte, error) {
	type alias LoginRequest
	aux := struct {
		*alias
		AuthMethod json.RawMessage `json:"auth_method,omitempty"`
	}{alias: (*alias)(m)}
	if m.AuthMethod != nil {
		var obj map[string]any
		switch v := m.AuthMethod.(type) {
		case *LoginRequestAuthMethodEmail:
			obj = map[string]any{"auth_type": "email", "email": v.Email}
		case *LoginRequestAuthMethodToken:
			obj = map[string]any{"auth_type": "token", "token": v.Token}
		case *LoginRequestAuthMethodSocial:
			obj = map[string]any{"auth_type": "social", "social": v.Social}
		}
		objBytes, err := json.Marshal(obj)
		if err != nil {
			return nil, err
		}
		aux.AuthMethod = objBytes
	}
	return json.Marshal(aux)
}

func (m *LoginRequest) UnmarshalJSON(data []byte) error {
	type alias LoginRequest
	aux := struct {
		*alias
		AuthMethod json.RawMessage `json:"auth_method,omitempty"`
	}{alias: (*alias)(m)}
	if err := json.Unmarshal(data, &aux); err != nil {
		return err
	}
	if len(aux.AuthMethod) > 0 {
		var disc struct {
			Tag string `json:"auth_type"`
		}
		if err := json.Unmarshal(aux.AuthMethod, &disc); err != nil {
			return err
		}
		switch disc.Tag {
		case "email":
			var v struct {
				Val *EmailAuth `json:"email"`
			}
			if err := json.Unmarshal(aux.AuthMethod, &v); err != nil {
				return err
			}
			m.AuthMethod = &LoginRequestAuthMethodEmail{Email: v.Val}
		case "token":
			var v struct {
				Val *TokenAuth `json:"token"`
			}
			if err := json.Unmarshal(aux.AuthMethod, &v); err != nil {
				return err
			}
			m.AuthMethod = &LoginRequestAuthMethodToken{Token: v.Val}
		case "social":
			var v struct {
				Val *SocialAuth `json:"social"`
			}
			if err := json.Unmarshal(aux.AuthMethod, &v); err != nil {
				return err
			}
			m.AuthMethod = &LoginRequestAuthMethodSocial{Social: v.Val}
		}
	}
	return nil
}

type LoginRequestAuthMethod interface {
	isLoginRequestAuthMethod()
}

type LoginRequestAuthMethodEmail struct {
	Email *EmailAuth `json:"email,omitempty"`
}

func (*LoginRequestAuthMethodEmail) isLoginRequestAuthMethod() {}

type LoginRequestAuthMethodToken struct {
	Token *TokenAuth `json:"token,omitempty"`
}

func (*LoginRequestAuthMethodToken) isLoginRequestAuthMethod() {}

type LoginRequestAuthMethodSocial struct {
	Social *SocialAuth `json:"social,omitempty"`
}

func (*LoginRequestAuthMethodSocial) isLoginRequestAuthMethod() {}

type LoginResponse struct {
	AccessToken  string `json:"access_token,omitempty"`
	RefreshToken string `json:"refresh_token,omitempty"`
	ExpiresIn    int64  `json:"expires_in,omitempty"`
	User         *User  `json:"user,omitempty"`
}